+++ src/utils.jl
@@ -1,17 +1,19 @@
-"""
-    load_templates!()
-
-Reloads the templates from PromptingTools.jl and the custom ones in LLMTextAnalysis.jl (saved in `templates/` folders).
-"""
-function load_templates!()
-    ## Load extra templates
-    PT.load_templates!() # refresh base templates
-    PT.load_templates!(joinpath(@__DIR__, "..", "templates"); remove_templates = false) # add our custom ones
-end
-
 "Counts number of unique elements in a vector"
 nunique(vect::AbstractVector) = length(unique(vect))
 sigmoid(x::Real) = 1.0 / (1.0 + exp(-x))
+function softmax(x::AbstractVector)
+    exp_ = exp.(x)
+    exp_ / sum(exp_)
+end
+function softmax(x::AbstractMatrix)
+    temp = exp.(x)
+    sum_ = sum(temp, dims = 2) |> vec
+    for j in axes(temp, 2)
+        # divide by sum of each row
+        temp[:, j] ./= sum_
+    end
+    return temp
+end
 
 """
     wrap_string(str::String,